crypto is now independent of dnssec::sign.
