+++ src/dnssec/mod.rs
@@ -2,8 +2,5 @@
 #[cfg(feature = "ring")]
 pub mod sign;
 
-// common requires crypto and crypto requires unstable-sign
-#[cfg(feature = "unstable-sign")]
 pub mod common;
-
 pub mod validator;
